requirements.txt
click==8.0.3
Flask==2.0.3
flask-wtf
itsdangerous==2.0.1
Jinja2==3.1.2
MarkupSafe==2.1.1
Werkzeug==2.2.2
opencv-python
numpy
matplotlib
pandas
Pillow
tensorflow
keras
ipython
psutil 
pyyaml 
tqdm 
seaborn
